pyproject.toml
[project]
name = "python-fs-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "e2b>=2.7.0",
    "e2b-code-interpreter>=2.3.0",
    "fastapi>=0.122.0",
    "langchain>=1.1.0",
    "langchain-anthropic>=1.2.0",
    "langchain-openai>=1.1.0",
    "langgraph-checkpoint-mongodb==0.2.2",
    "langgraph-store-mongodb>=0.1.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
